Refactor : add separate component for entity version.
