+++ catalog-rest-service/src/main/resources/ui/src/pages/EntityVersionPage/EntityVersionPage.component.tsx
@@ -1,7 +1,5 @@
 import { AxiosError, AxiosResponse } from 'axios';
-import classNames from 'classnames';
-import { ColumnJoins } from 'Models';
-import React, { useEffect, useState } from 'react';
+import React, { FunctionComponent, useEffect, useState } from 'react';
 import { useHistory, useParams } from 'react-router-dom';
 import { getDatabase } from '../../axiosAPIs/databaseAPI';
 import { getServiceById } from '../../axiosAPIs/serviceAPI';
@@ -10,20 +8,14 @@ import {
   getTableVersion,
   getTableVersions,
 } from '../../axiosAPIs/tableAPI';
-import Description from '../../components/common/description/Description';
-import EntityPageInfo from '../../components/common/entityPageInfo/EntityPageInfo';
-import TabsPane from '../../components/common/TabsPane/TabsPane';
 import { TitleBreadcrumbProps } from '../../components/common/title-breadcrumb/title-breadcrumb.interface';
-import PageContainer from '../../components/containers/PageContainer';
-import EntityVersionTimeLine from '../../components/EntityVersionTimeLine/EntityVersionTimeLine';
+import DatasetVersion from '../../components/DatasetVersion/DatasetVersion.component';
 import Loader from '../../components/Loader/Loader';
-import SchemaTab from '../../components/SchemaTab/SchemaTab.component';
 import {
   getDatabaseDetailsPath,
   getDatasetDetailsPath,
   getDatasetVersionPath,
   getServiceDetailsPath,
-  getTeamDetailsPath,
 } from '../../constants/constants';
 import { Table } from '../../generated/entity/data/table';
 import { EntityHistory } from '../../generated/type/entityHistory';
@@ -31,8 +23,7 @@ import useToastContext from '../../hooks/useToastContext';
 import { getPartialNameFromFQN } from '../../utils/CommonUtils';
 import { serviceTypeLogo } from '../../utils/ServiceUtils';
 import { getOwnerFromId, getTierFromTableTags } from '../../utils/TableUtils';
-import { getTableTags } from '../../utils/TagsUtils';
-const EntityVersionPage = () => {
+const EntityVersionPage: FunctionComponent = () => {
   const history = useHistory();
   const showToast = useToastContext();
   const [tier, setTier] = useState<string>();
@@ -51,38 +42,7 @@ const EntityVersionPage = () => {
   const [slashedTableName, setSlashedTableName] = useState<
     TitleBreadcrumbProps['titleLinks']
   >([]);
-  const [isDrawerOpen, setIsDrawerOpen] = useState<boolean>(true);
 
-  const extraInfo = [
-    {
-      key: 'Owner',
-      value:
-        owner?.type === 'team'
-          ? getTeamDetailsPath(owner?.name || '')
-          : owner?.name || '',
-      placeholderText: owner?.displayName || '',
-      isLink: owner?.type === 'team',
-      openInNewTab: false,
-    },
-    { key: 'Tier', value: tier ? tier.split('.')[1] : '' },
-  ];
-
-  const tabs = [
-    {
-      name: 'Schema',
-      icon: {
-        alt: 'schema',
-        name: 'icon-schema',
-        title: 'Schema',
-      },
-      isProtected: false,
-      position: 1,
-    },
-  ];
-
-  const versionDrawerHandler = () => {
-    setIsDrawerOpen((prevState) => !prevState);
-  };
   const backHandler = () => {
     history.push(getDatasetDetailsPath(datasetFQN));
   };
@@ -193,64 +153,24 @@ const EntityVersionPage = () => {
   }, [version]);
 
   return (
-    <PageContainer>
-      <div
-        className={classNames(
-          'tw-px-4 tw-w-full tw-h-full tw-flex tw-flex-col tw-relative'
-        )}>
-        {isVersionLoading ? (
-          <Loader />
-        ) : (
-          <div className={classNames({ 'version-data': isDrawerOpen })}>
-            <EntityPageInfo
-              entityName={currentVersionData.name}
-              extraInfo={extraInfo}
-              followersList={[]}
-              tags={getTableTags(currentVersionData.columns || [])}
-              tier={tier || ''}
-              titleLinks={slashedTableName}
-              version={version}
-              versionHandler={versionDrawerHandler}
-            />
-            <div className="tw-mt-1 tw-flex tw-flex-col tw-flex-grow ">
-              <TabsPane activeTab={1} className="tw-flex-initial" tabs={tabs} />
-              <div className="tw-bg-white tw-flex-grow">
-                <div className="tw-grid tw-grid-cols-4 tw-gap-4 tw-w-full tw-mt-4 ">
-                  <div className="tw-col-span-full">
-                    <Description
-                      isReadOnly
-                      description={currentVersionData.description || ''}
-                    />
-                  </div>
-
-                  <div className="tw-col-span-full">
-                    <SchemaTab
-                      isReadOnly
-                      columnName={getPartialNameFromFQN(
-                        datasetFQN,
-                        ['column'],
-                        '.'
-                      )}
-                      columns={currentVersionData.columns}
-                      joins={currentVersionData.joins as ColumnJoins[]}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
-        )}
-
-        <EntityVersionTimeLine
-          currentVersion={version}
-          isLoading={isLoading}
-          show={isDrawerOpen}
+    <>
+      {isLoading ? (
+        <Loader />
+      ) : (
+        <DatasetVersion
+          backHandler={backHandler}
+          currentVersionData={currentVersionData}
+          datasetFQN={datasetFQN}
+          isVersionLoading={isVersionLoading}
+          owner={owner}
+          slashedTableName={slashedTableName}
+          tier={tier as string}
+          version={version}
           versionHandler={versionHandler}
           versionList={versionList}
-          onBack={backHandler}
         />
-      </div>
-    </PageContainer>
+      )}
+    </>
   );
 };
